feat: move UserProfileBookmarksFeedFilter and VM to commons
Move the bookmarks feed filter and view model to the commons module,
using IAccount and ICacheProvider interfaces. The amethyst-side VM now
extends the commons version, passing LocalCache as the cache provider.
